refactor(procmond): split event_bus_connector.rs into event_bus_connector/ submodules
Pure mechanical module extraction with no logic or public-API changes:
- event_bus_connector/error.rs: EventBusConnectorError + EventBusConnectorResult
- event_bus_connector/signal.rs: BackpressureSignal + ProcessEventType (methods pub(super))
- event_bus_connector/buffered_event.rs: BufferedEvent struct + impl (pub(super))
- event_bus_connector/tests.rs: full unit test module
- event_bus_connector/mod.rs: EventBusConnector struct + impl, consts, re-exports

All 58 event_bus_connector unit tests pass; build/clippy/fmt clean.
